Some Changes
you don't need 3 different sets for one loop

changed some names and numberings

(not sure if this is for my own fork or the branch or what)
