+++ src/main/java/edu/rpi/legup/puzzle/sudoku/SudokuCellFactory.java
@@ -16,6 +16,7 @@ public class SudokuCellFactory extends ElementFactory {
      *
      * @param node  node that represents the puzzleElement
      * @param board board to add the newly created cell
+     * @param board board to add the newly created cell
      * @return newly created cell from the xml document Node
      * @throws InvalidFileFormatException if file is invalid
      */
