+++ src/main/java/edu/rpi/legup/puzzle/sudoku/rules/LastCellForNumberDirectRule.java
@@ -12,7 +12,7 @@ import java.util.Set;
 
 public class LastCellForNumberDirectRule extends DirectRule {
     public LastCellForNumberDirectRule() {
-        super("SUDO-BASC-0002", "Last Cell for Number",
+        super("SUDO-BASC-0001", "Last Cell for Number",
                 "This is the only cell open in its group for some number.",
                 "edu/rpi/legup/images/sudoku/forcedByElimination.png");
     }
